[DEV] programs: Use int8_t to test flag and shift instructions

+++ programs/apps/hello/main.c
@@ -4,7 +4,7 @@
 
 int main(void) {
   const char* hello = "hello";
-  uint16_t i;
+  int8_t i;
 
   for (i = 0; i < 6; i++)
     *(SERIAL + i) = hello[i];
